Moved all environments into experiment subfolders
Cleans up the organization on the Xopt-GEECS side, although all the symlinks are still in the same place in the conda env side
