+++ Xopt-GEECS/badger-plugins/environments/examples/example_single_camera_median/configs.yaml
