+++ Xopt-GEECS/badger-plugins/environments/examples/example_single_camera_target/configs.yaml
