+++ Xopt-GEECS/badger-plugins/environments/examples/example_two_camera_math/__init__.py
@@ -1,6 +1,9 @@
 import numpy as np
 from typing import List
-from ..parent_geecs_env import Environment as geecs_Environment
+try:  # The top import works in Badger with symlinks, the bottom import lets PyCharm know what is being imported
+    from ..base_geecs_env import Environment as geecs_Environment
+except ImportError:
+    from ...geecs_general.base_geecs_env import Environment as geecs_Environment
 
 
 class Environment(geecs_Environment):
