+++ Xopt-GEECS/badger-plugins/environments/examples/example_two_camera_math/configs.yaml
